Add support for testing content scripts with Cypress
Before this commit, content scripts couldn't be tested with cypress, because they wouldn't get injected into the frame controlled by Cypress. The content scripts were instead injected in the main frame of the page used by Cypress to contain the iframe with the page being tested. To get the content scripts to be injected into the correct frame during testing, I added a webpack configuration that could search and replace text in the background script. The background script uses helper functions to inject the content scripts. I modified the helper functions to take a  argument. Now, I explicitly pass the correct frame id for the frame I inject. However, I cannot simply enter fixed frame id when testing, because the frame id in Cypress always changes on reload. Instead, I replace the guard clause used in production to filter out  events from child frames to a guard clause that filters out  events from frames that are not the page I'm testing. To accomplish that I sue the  to check the origin of the  event. If it is the page I'm testing, then it passes the filter. I use a Webpack config for testing to replace the guard clause. Note, I haven't generalized the url test, right now it works only for one page I explicitly code in.