Attempt to support sending messages from Cypress controlled page to extension through chrome.runtime.sendMessage
This didn't work because chrome.runtime is set to undefined by chrome when the page isn't loaded with https, even if the hostname is added to the externally_connectable key in the manifest.json. I tried to make Cypress load the page with https by configuring the server that serves test pages to run with https, however Cypress uses a proxy that adds it's own certificate which chrome doesn't accept. The result is chrome.runtime is still undefined and I can't send a message from Cypress to extension.