Add command to reset extension state before each test
Woo! Finally did it! The convential way of communicating with an extension from a webpage is through chrome.runtime.sendMessage. But that doesn't work in Cypress since Cypress doesn't run https securely (Chrome doesn't accept the certficate given by the Cypress proxy as valid), so chrome.runtime is set to undefined (chrome.runtime is available only on https). This commit adds a workaround. It works by adding a message relay extension to the browser controlled by
Cypress. The relay extension listens for an event on the document and then sends a message to the chrome extension I'm testing. I trigger the event from within a frame controlled by Cypress using a cy command I defined. Since the message relay is running on all frames, the event triggers it's listener callback and the message is relayed to the extension. This all works because extension content scripts always have access to chrome.runtime.sendMessage and to the DOM of the frame they're attached to.