Complete refactoring to message-oriented content scripts!
Before this refactoring, the background script would run the specific content scripts on a tab when they were needed. Kind of like RPC in networking. The problem is that design is very brittle and would make upcoming features difficult to implement. After the refactoring, the background script only injects one content script on any given page, the user-view-controller (better name might be found eventually). That content script is like it's own independent machine that is responsible for the user view (altering the pages the browser user visits). The background script sends messages to that "machine" and it can do as it wishes when it receives those messages. This effectively isolates all design decisions made to implement the content scripts from the background script. The background scripts has no knowledge of the content scripts. It simply sends messages there.