Extract onMessageExternal listener to message relay extension
The onMessageExternal listener in background.js is only used in testing for the message relay extension (pluged in to Cypress-controled browser) to relay messages from Cypress to the extension being tested. It allows Cypress to control the state of the extension programatically. Since it is only used in testing, I extracted it to simplify the background.js file. Now it is imported by webpack when building the extension for testing.